+++ src/prompts/component-type.ts
@@ -1,6 +1,6 @@
 import { window } from 'vscode';
 
-type ComponentType = 'class' | 'functional';
+type ComponentType = 'class' | 'class-pure' | 'functional' | 'functional-memo';
 type QuickPickComponentType = {
     type: ComponentType;
     label: string;
@@ -15,8 +15,16 @@ export default async (): Promise<ComponentType> => {
                     label: `类 (Class)`
                 },
                 {
+                    type: 'class-pure',
+                    label: `类 (Class) - PureComponent`
+                },
+                {
                     type: 'functional',
                     label: `函数组件 (Functional Component)`
+                },
+                {
+                    type: 'functional-memo',
+                    label: `函数组件 (Functional Component) - memo`
                 }
             ],
             {