+++ templates/class-typescript/index.tsx
@@ -15,7 +15,7 @@ interface ComponentProps {
     /* NEED_CHANGE_ALL_EXTEND_OPTIONS_IN_COMMENTED */
     styles
 })
-class NEED_CHANGE_COMPONENT_NAME extends React.Component<
+class NEED_CHANGE_COMPONENT_NAME extends React.NEED_CHANGE_COMPONENT_TYPE<
     ExtendedProps & ComponentProps
 > {
     render(): React.ReactNode {